fix(error-overlay): improve a11y, minor refactors
