+++ packages/next/src/client/components/react-dev-overlay/internal/container/Errors.tsx
@@ -25,7 +25,7 @@ import { VersionStalenessInfo } from '../components/VersionStalenessInfo'
 import type { VersionInfo } from '../../../../../server/dev/parse-version-info'
 import { getErrorSource } from '../../../../../shared/lib/error-source'
 import { HotlinkedText } from '../components/hot-linked-text'
-import { PseudoHtmlDiff } from './RuntimeError/component-stack-pseudo-html'
+import { PseudoHtmlDiff } from './RuntimeError/ComponentStackPseudoHtml'
 import {
   isHtmlTagsWarning,
   type HydrationErrorState,
